Made most time variables 64 bits. Added unittest-periodic.
git-svn-id: svn://graphite/diminuto/trunk@35 5dfbf39a-1452-0410-8a97-9e58fedde8e3

+++ Diminuto/diminuto_periodic.h
@@ -25,6 +25,6 @@
  * @return the number of microseconds remaining in the prior timer if
  * a timer was already running.
  */
-extern uint32_t diminuto_periodic(uint32_t microseconds);
+extern uint64_t diminuto_periodic(uint64_t microseconds);
 
 #endif
